fix/Move the nest-asyncio apply into an init hook
We're seeing an issue where the nest_asyncio.apply() workaround for nested loops is breaking when
we're dealing with a `uvloop`. We need to investigate further, or remove the bandaid solution. In
the meantime, we can unblock a simple import of the client in these environments by doing the apply
only when the hook is run.