fix: Do not throw error when in a uvloop context (#142)
PDF page splitting uses asyncio but the SDK is not async. Therefore, we
had to manage our own event loop, which can lead to issues in other
event loop contexts. Uvloop is one context that does not allow us to use
nested event loops. When we find ourselves in a uvloop.Loop, we have to
fallback to non splitting mode. #135 will make the whole SDK async so we
don't have to hack this.

Closes #133