fix: handle RuntimeError when checking if we can use pdf split mode (#145)
Certain environments throw a `RuntimeError: There is no current event
loop in thread 'asyncio_0'` when we try to check if we're able to use
the splitting code. In this case, return False so we can fallback to non
split mode.

Once this is merged, I can regenerate the client to publish 0.25.4.