fix: Address mypy typing errors in v2 SDK
Update all typing that used requests to take httpx classes instead. Logic changes should be minimal,
this is mostly to change type hints where a `httpx.Response` is used instead of a
`requests.Response`, etc.

Add mypy to `make lint` so that we can cath these errors before merging. The publish job runs a full
linter suite, and these changes made it to main but broke the publish job.