fix/ensure that split pdf requests are retried
We discovered that when a pdf is split into smaller chunks, those requests are not being
retried. Now that we have `allow_failed=False`, this results in the whole document failing as soon
as any of the child requests hit a transient error. The fix is to reuse the `utils.Retry` logic that
the main code path uses. Copying the retry config in the hook logic is not great, and we can work
with Speakeasy to make the internal logic more modular. But for now, this will address the current
failures while we work on a better implementation.

Testing:
See the added unit test. The existing retry logic works for the final split page, everything else
needs to use the new logic. To test this, I mocked a response from the server to return 502 for a
low `starting_page_number`, which we know will have to be handled by the hooks.