fix/ensure that split pdf requests are retried (#160)
We discovered that when a pdf is split into smaller chunks, those
requests are not being retried. Now that we have `allow_failed=False`,
this results in the whole document failing as soon as any of the child
requests hit a transient error. The fix is to reuse the `utils.Retry`
logic that the main code path uses. Copying the retry config in the hook
logic is not great, and we can work with Speakeasy to make the internal
logic more modular so we can reuse more. But for now, this will address
the current failures while we work on a better implementation.

Testing:
See the added unit test. The existing retry logic works for the final
split page, everything else needs to use the new logic. To test this, I
mocked a response from the server to return 502 for a low
`starting_page_number`, which we know will have to be handled by the
hooks.

Other changes:
Remove the "Not splitting" log. When the final split page is retried, it
triggers all the hooks again. We need to force `split_pdf_page=False` in
this request, and we don't need additional logging when this code is hit
again.