Add the no-op client mock for httpx.AsyncClient
The hooks do not give us access to the AsyncClient, which breaks the mock no-op request in the
splitting hook. For now, let's add it as a param to the sdk_init hook. This means we'll have to
bring this code back whenever the SDK regenerates. At least, until we have an update from Speakeasy
for this.