fix: Address some issues in the split pdf logic (#170)
This is a copy of [the
fixes](https://github.com/Unstructured-IO/unstructured-python-client/pull/165)
that went into the release/0.25.x branch, updated slightly for the v2
SDK. The main branch is already retrying split requests, but we do not
have an updated httpx timeout value, and the retries are not logged.

# To verify
Same as the linked pr, we can mock arbitrary errors from the server and
verify the behavior.

Apply this patch to your local unstructured-api, and do `make
run-web-app`. Whenever we send a request starting at page 3, the server
will sleep for a bit, and then return a 502 error.

```
diff --git a/prepline_general/api/general.py b/prepline_general/api/general.py
index 2ca56270..0624076c 100644
*** a/unstructured-api/prepline_general/api/general.py
--- b/unstructured-api/prepline_general/api/general.py
***************
*** 812,817 ****
--- 814,822 ----
      )
  
  
+ global num_bounces
+ num_bounces = 2
+ 
  @elasticapm.capture_span()
  @router.post(
      "/general/v0/general",
***************
*** 835,840 ****
--- 840,854 ----
      # For new parameters - add them in models/form_params.py
      form_params: GeneralFormParams = Depends(GeneralFormParams.as_form),
  ):
+ 
+     global num_bounces
+     page_num = form_params.starting_page_number or 1
+     if num_bounces > 0 and page_num == 3:
+         num_bounces -= 1
+         time.sleep(10)
+         raise HTTPException(status_code=502, detail="BOUNCE")
+ 
      # -- must have a valid API key --
      if api_key_env := os.environ.get("UNSTRUCTURED_API_KEY"):
          api_key = request.headers.get("unstructured-api-key")
```

Use the main branch to make a client request using layout-parser-paper,
which should default to 2 page splits, with the second split starting at
page 3. Because the httpx timeout is not specified, this split will
raise a ReadTimeout error, and get stuck in a retry loop. This is not
logged, so it will appear to be a client hang.

```
from unstructured_client import UnstructuredClient
from unstructured_client.models import shared, operations

filename = "_sample_docs/layout-parser-paper.pdf"

s = UnstructuredClient(
    server_url="http://localhost:5000",
)

with open(filename, "rb") as f:
    files=shared.Files(
        content=f.read(),
        file_name=filename,
    )

req = operations.PartitionRequest(
    shared.PartitionParameters(
        files=files,
        strategy="fast",
        split_pdf_page=True,
    ),
)

resp = s.general.partition(req)

if num_elements := len(resp.elements):
    print(f"Succeeded with {num_elements}")
```

Finally, make the same request against this branch. The bad split should
be retried for the allotted 2 times, with logging, and then the doc will
succeed.