fix: Allow split page logic to process files concurrently (#175)
# The Issue

`split_pdf_hook.py` does not support multiple concurrent files. This is
because we store the split request tasks in
`self.coroutines_to_execute[operation_id]`, where `operation_id` is just
the string "partition". Therefore, if we send two concurrent docs using
the same SDK, they'll both try to await the same list of coroutines.
This could result in interleaved results, but mostly it breaks with
`RuntimeError: coroutine is being awaited already`, as the second
request gets ready to await its requests. This will block anyone trying
to use the new `partition_async` to fan out their pdfs.

Note that the js/ts client also has this issue.

# The fix

We need to use an actual id to index into `coroutines_to_execute`. In
`before_request`, let's make a uuid and build up the list of coroutines
for this doc. We need to pass this id to `after_success` in order to
retrieve the results, so we can set it as a header on our "dummy"
request that's returned to the SDK.

# Testing

See the new integration test. We can verify this by sending two docs
serially, and then with `asyncio.gather`, and confirm that the results
are the same.