chore: Add min version requirement for `cryptography` (#179)
Bring the min version to the main branch - this is done for the 0.25.x
release
[here](https://github.com/Unstructured-IO/unstructured-python-client/pull/176).

Also, remove jupyter as a dev dependency. You're free to install it for
local work, but we don't need to pull it on every `poetry install`. Case
in point, it was causing an issue in the [generate
action](https://github.com/Unstructured-IO/unstructured-python-client/actions/runs/10947329413/job/30395826215).
The jupyter install was trying to pull in `psutil`, which needed to be
built from source. However, our generate environment doesn't have all
the build tools set up, so let's remove the need for `psutil` entirely.