fix: Use configured server_url for split page dummy request (#192)
Fixes #191. Due to a limitation in our templated codebase, we need our
split page logic to return a request that's just going to give a 200
response. Then, we jump back into the AfterSuccessHook and await all of
the pdf splits. Hopefully we can clean this up soon to avoid having an
extra call at all, but for now let's make sure the GET /general/docs is
done on the correct base url.