fix/Address server_url breaking change (#226)
In order to add the platform-api functions, we now need to handle two
openapi specs. When these are merged, we get a breaking change with the
`server_url` arg:

```
# The passed url is ignored and we go to our default Serverless URL. 
# You suddenly get an invalid api key error
s = UnstructuredClient(server_url="my_own_url")
s.general.partition()

#  This does the right thing
s = UnstructuredClient()
s.general.partition(server_url="my_own_url")
```

We published the 0.30.0 client with this change, but let's do the right
thing before it breaks a whole lot of users. For now, let's directly
edit the endpoint functions and make sure the client level url is still
accepted. We can work with Speakeasy to roll this back into the auto
generated SDK.

## Changes
- Add a helper function to our `clean_server_url` hook file to select
the right url from three options: the endpoint, the client, or the
default from the config
- Call this helper from every generated method (plus async versions)
- Add a `test_server_urls.py` to test every URL case against every
endpoint. The user can pass the URL in either location, and the SDK will
strip the path if it's given, for backwards compat.
- Add the endpoint files to `.genignore`. This will unblock the generate
action for now, and our test suite will tell us if these files have
gotten stale.

Other:
- Update `workflow.yaml` to reorder the openapi specs. Both specs
include `HttpValidationError`, and the later spec will overwrite the
duplicate. This schema changed a bit in Serverless, so some tests are
failing on main. Note that the `make client-generate-sdk` does the right
thing, but this got reverted when the github generate action ran again.

## Testing
- Check out this branch and verify that your old SDK code with a custom
URL is still working