chore: Clean up some connector schemas in the overlay file (#247)
This is to hold us over until these are updated on the backend. Add
overlay config to remove existing schemas and recreate with a new name.
This is a really ugly bandaid, we don't want to store actual schemas in
this file for any longer than we have to.

Some tests are broken because we merged the schemas that we don't want,
this should fix that on the next generate job.

Changes:

Onedrive -> OneDrive
Kafka-cloud -> KafkaCloud
gcs -> GCS
Google_drive -> GoogleDrive