chore: Manually regenerate SDK (#299)
We have a generate failure because we've pulled `general.py` out for
some custom changes. (See
https://github.com/Unstructured-IO/unstructured-python-client/pull/270)
This will hit the occasional bump as new code is added to this file
upstream. The steps for fixing this:

- Remove `general.py` from .genignore
- Run speakeasy generate
- Stage and commit the autogenerated changes, working around our custom
code
- Commit without the `.genignore` change

I set `gen.yaml` to version 0.41.0. This will cause the next generate
job to propagate the new version and publish it.