feat: Enable arbitrary input for ConnectorType enums
Use the openapi overlay file to add `x-speakeasy-unknown-values` to the models for
`SourceConnectorType` and `DestinationConnectorType`. This allows the client to send any string
outside of the enum definitions. This provides forward compatibility in the client, we can create
sources with types that don't exist yet, without requiring a new client version.

Example:

This throws an error that the type is not in our enum. Now, it just warns that this is an unknown
value, but it's sent to the server anyway.

```
res = unstructured_client.sources.create_source(
    request={
        "create_source_connector": {
            "name": "My fancy new source",
            "type": "future_source_type",
            "config": {
                ...
                ...
            }
        }
    }
)
'''

The change will take effect when the client regenerates and uses the new overlay config. By setting
`gen.yaml` to `0.42.2`, the new client will propagate this version change.